라우터 수정 전
